Compare positions in tryTypedApply by focus.
Synthetic trees usually get offset positions, even with
range positions enabled. The comparison previously used
by `errorInResult` did not consider the error issued by
`AdaptTypeError` on an `ApplyToImplicitArgs` to be part
of the result expression, meaning that an implicit view
wouldn't be sought on a second try.

Fixes scala/bug#10643.