Intrinsify Enumeration's Value method.
Java reflection's field order is undefined and potentially
variable between runs; defining a `val` that refers to one
of the enumeration's `Value`s can cause wrong names to get
put in the names map at runtime. We can do the computation
at compile time with a fast-tracked macro instead.

This fixes scala/bug#5462.

Also, it should help to simplify scala-js, which evidently
special-cases this class pretty hard to compensate for the
lack of Java reflection in Javascript.

Because `Enumeration` is used in the library and compiler,
the removal of the non-macro `Value` methods isn't able to
happen until 2.13.x is reSTARRed. Therefore, this includes
a few temporary changes:

- `Value`'s old implementation moves to `_Value`, and gets
  the `protected[scala]` designation. All users of `Value`
  in the compiler and library temporarily call this method
  instead, until the intrinsic is in STARR.
- A few tests that depend on `scala-xml` get suspended for
  now, until there's a version that uses the new intrinsic
  method. I've somewhat whimsically replaced the test code
  with a `println` reproducing the check file, in the hope
  that, when the test is re-enabled, the git history won't
  vanish into the aether.